Moved PatchStarCard to player patch section (as it is used for playable character patch set)
